# lib/api-zod/src/generated/types/architecturalProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProgramSpace } from "./programSpace";

export interface ArchitecturalProgram {
  totalArea: number;
  spaces: ProgramSpace[];
  designPrinciples: string[];
  materialPalette: string[];
  sustainabilityStrategies: string[];
  estimatedBudget?: string;
  timeline?: string;
  styleDirection: string;
}