# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyProjectType } from "./createProjectBodyProjectType";

export interface CreateProjectBody {
  name: string;
  description?: string;
  projectType: CreateProjectBodyProjectType;
}