# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryProjectsByStatus } from "./dashboardSummaryProjectsByStatus";
import type { DashboardSummaryProjectsByType } from "./dashboardSummaryProjectsByType";

export interface DashboardSummary {
  totalProjects: number;
  projectsByStatus: DashboardSummaryProjectsByStatus;
  projectsByType: DashboardSummaryProjectsByType;
  totalImages: number;
  recentActivity: string[];
}