# lib/api-zod/src/generated/types/generateImageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateImageBodyImageType } from "./generateImageBodyImageType";

export interface GenerateImageBody {
  imageType: GenerateImageBodyImageType;
  title: string;
  customPrompt?: string;
}