# lib/api-zod/src/generated/types/programSpace.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProgramSpacePriority } from "./programSpacePriority";

export interface ProgramSpace {
  name: string;
  area: number;
  quantity: number;
  description?: string;
  priority: ProgramSpacePriority;
}