# lib/api-zod/src/generated/types/projectImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectImageImageType } from "./projectImageImageType";

export interface ProjectImage {
  id: number;
  projectId: number;
  imageType: ProjectImageImageType;
  prompt: string;
  imageUrl: string;
  title: string;
  createdAt: string;
}