# lib/api-zod/src/generated/types/questionnaire.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionnaireAnswers } from "./questionnaireAnswers";

export interface Questionnaire {
  answers: QuestionnaireAnswers;
}