# lib/api-zod/src/generated/types/siteAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archi Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteAnalysis {
  climate: string;
  soilType: string;
  sunExposure: string;
  windDirection: string;
  topography: string;
  surroundingContext: string;
  opportunities: string[];
  constraints: string[];
  recommendations: string[];
  sustainabilityScore: number;
  accessibilityNotes?: string;
  zoningInfo?: string;
}